Start of the HierarchicalDisplay. Using a new gwt-dragdrop library.
git-svn-id: file:///a/svnrepos/hippo/code/trunk/HippoTest@382 b8d5060c-ad15-0410-9d10-96fc93d4f2a4

+++ src/main/java/com/aavu/client/domain/dto/FullTopicIdentifier.java
@@ -7,8 +7,8 @@ import com.google.gwt.user.client.rpc.IsSerializable;
 
 public class FullTopicIdentifier extends DatedTopicIdentifier implements IsSerializable {
 
-	private double latitudeOnIsland;
-	private double longitudeOnIsland;
+	private int latitudeOnIsland;
+	private int longitudeOnIsland;
 	
 	public FullTopicIdentifier(){}
 	
@@ -29,19 +29,19 @@ public class FullTopicIdentifier extends DatedTopicIdentifier implements IsSeria
 		this.longitudeOnIsland = conn.getLongitude();
 	}
 
-	public double getLatitudeOnIsland() {
+	public int getLatitudeOnIsland() {
 		return latitudeOnIsland;
 	}
 
-	public void setLatitudeOnIsland(double latitude) {
+	public void setLatitudeOnIsland(int latitude) {
 		this.latitudeOnIsland = latitude;
 	}
 
-	public double getLongitudeOnIsland() {
+	public int getLongitudeOnIsland() {
 		return longitudeOnIsland;
 	}
 
-	public void setLongitudeOnIsland(double d) {
+	public void setLongitudeOnIsland(int d) {
 		this.longitudeOnIsland = d;
 	}
